refactor(suite): move send form thunks to suite-common
