--- packages/suite/src/actions/wallet/send/constants.ts
@@ -1,1 +1,0 @@
-export const MODULE_PREFIX = '@send';
